requirements.txt
python-docx
python-pptx
prompty[azure]
openai
azure-identity